\name{gemm.plot}
\alias{gemm.plot}
\title{plot method for gemm}
\usage{
  gemm.plot(x, ...)
}
\arguments{
  \item{x}{"gemm" class object to be used for plotting.
  Note that plots differ based on whether
  \code{check.convergence} was called for the initial gemm
  fitting.}

  \item{...}{Other arguments.}
}
\description{
  This function is the plot method for a gemm object.
}

