\name{gemm.print}
\alias{gemm.print}
\title{Print method for gemm}
\usage{
  gemm.print(x, ...)
}
\arguments{
  \item{x}{gemm object to be referenced.}

  \item{...}{Other arguments.}
}
\description{
  This function is the print method for gemm. It returns
  the most-referenced parts of a gem object: formula call,
  coefficients, and fit indices.
}

