\name{gemmFitRcppI}
\alias{gemmFitRcppI}
\title{gemmFitRcppI}
\usage{
  gemmFitRcppI(n, betas, data, p, kCor, getR)
}
\arguments{
  \item{n}{Number of observations in dataset.}

  \item{betas}{Matrix of random weights.}

  \item{data}{Dataset constructed based on original
  \code{\link{gemm.formula}} call. First column is
  criterion variable.}

  \item{p}{Number of potential predictors in data.}

  \item{kCor}{Appropriate penalty term for given vector of
  weights. Will always equal number of non-zero weights
  unless interactions are included.}

  \item{getR}{Whether or not to return Pearson's r. To save
  computation time, r is only returned on the last
  generation of each chain.}
}
\value{
  fit List of named vectors for BICtau, BICr, tau, and r.
}
\description{
  This function is run each generation to calculate
  Kendall's tau-based fit between weighted predictors,
  (weights generated by \code{\link{genAlg}}), and
  criterion. Calls a number of Rcpp functions that are
  hidden from R.
}

