\name{list2gemm}
\alias{list2gemm}
\title{Collapse a uniform list of \code\link{{gemm}} objects.}
\usage{
  list2gemm(gemm.list)
}
\arguments{
  \item{gemm.list}{a uniform list of \code{gemm} objects, as generated by a parallelizing function.}
}
\description{
  \code{list2gemm} is used to generate a normal \code{\link{gemm}} object from a list of gemm objects with identical input options and n.chains = 1.
}
\details{
  Will generally only be used in conjunction with a parallel processing package, otherwise equivalent to specifying higher n.chains argument in \link{\code{gemm}}.
}
\value{
  Output is a standard gemm object.
}
\seealso{
  \code{\link{gemm}}
}
\examples{
    library(gemmR)
    data(mtcars)
    fit <- list()
    for (i in 1:3) {
      fit[[i]] <- gemm(mpg ~ disp + cyl, data = mtcars, n.chains = 1, n.gens = 3, n.beta = 200)
    }
    gemm.model <- list2gemm(fit)
    summary(gemm.model)
  }
}
\keyword{
  ordinal, regression
}
